major changes
user can now setup game, roll for players, and choose to either buy property or not buy it. application is ugly and barely functional. more changes coming soon, like end of turn actions. and hopefully a better looking ui system that will allow players to see on the gameBox what their options are